refactor, add weekly aggregation
